backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
requests==2.31.0
numpy==1.24.3
scikit-learn==1.3.0
python-decouple==3.8
django-cors-headers==4.3.1
